+++ openbb_terminal/miscellaneous/i18n/en.yml
@@ -31,7 +31,7 @@ en:
   account/list: List available routines
   account/upload: Upload routine
   account/download: Download routine
-  account/delete: Delete routine
+  account/delete: Delete routine stored in cloud
   keys/_source: Source
   keys/_keys_: This menu allows you to set your own API keys. Type 'about' for more information
   keys/mykeys: get your defined keys
