+++ tests/openbb_terminal/account/test_account_controller.py
@@ -308,7 +308,7 @@ def test_call_logout(mocker, test_user):
     path_controller = "openbb_terminal.account.account_controller"
 
     mocker.patch(
-        target="openbb_terminal.account.account_controller.get_current_user",
+        target=path_controller + ".get_current_user",
         return_value=test_user,
     )
 
@@ -333,7 +333,7 @@ def test_call_clear(mocker, input_value, test_user):
     controller = account_controller.AccountController(queue=None)
     path_controller = "openbb_terminal.account.account_controller"
     mocker.patch(
-        target="openbb_terminal.account.account_controller.get_current_user",
+        target=path_controller + ".get_current_user",
         return_value=test_user,
     )
     mock_input = mocker.patch(
@@ -360,7 +360,7 @@ def test_call_list(mocker, test_user):
     path_controller = "openbb_terminal.account.account_controller"
 
     mocker.patch(
-        target="openbb_terminal.account.account_controller.get_current_user",
+        target=path_controller + ".get_current_user",
         return_value=test_user,
     )
 
@@ -386,12 +386,12 @@ def test_call_upload(mocker, test_user):
     path_controller = "openbb_terminal.account.account_controller"
 
     mocker.patch(
-        target="openbb_terminal.account.account_controller.get_current_user",
+        target=path_controller + ".get_current_user",
         return_value=test_user,
     )
 
-    mock_get_routine = mocker.patch(
-        target=f"{path_controller}.Local.get_routine",
+    mock_read_routine = mocker.patch(
+        target=f"{path_controller}.read_routine",
         return_value="do something",
     )
     mock_upload_routine = mocker.patch(
@@ -409,7 +409,7 @@ def test_call_upload(mocker, test_user):
         ]
     )
 
-    mock_get_routine.assert_called_once_with(file_name="script1.openbb")
+    mock_read_routine.assert_called_once_with(file_name="script1.openbb")
     mock_upload_routine.assert_called_once_with(
         auth_header="Bearer 123",
         name="script1",
@@ -424,7 +424,7 @@ def test_call_download(mocker, test_user):
     path_controller = "openbb_terminal.account.account_controller"
 
     mocker.patch(
-        target="openbb_terminal.account.account_controller.get_current_user",
+        target=path_controller + ".get_current_user",
         return_value=test_user,
     )
 
@@ -443,7 +443,7 @@ def test_call_download(mocker, test_user):
     )
     mock_download_routine.return_value = response
     mock_save_routine = mocker.patch(
-        target=f"{path_controller}.Local.save_routine",
+        target=f"{path_controller}.save_routine",
         return_value="path_to_file",
     )
 
@@ -512,7 +512,7 @@ def test_call_generate(mocker, monkeypatch, test_user):
     ).encode("utf-8")
 
     mocker.patch(
-        target="openbb_terminal.account.account_controller.get_current_user",
+        target=path_controller + ".get_current_user",
         return_value=test_user,
     )
 
@@ -544,7 +544,7 @@ def test_call_show(mocker, test_user):
     ).encode("utf-8")
 
     mocker.patch(
-        target="openbb_terminal.account.account_controller.get_current_user",
+        target=path_controller + ".get_current_user",
         return_value=test_user,
     )
 
@@ -561,7 +561,7 @@ def test_call_revoke(mocker, monkeypatch, test_user):
     path_controller = "openbb_terminal.account.account_controller"
 
     mocker.patch(
-        target="openbb_terminal.account.account_controller.get_current_user",
+        target=path_controller + ".get_current_user",
         return_value=test_user,
     )
