+++ tests/openbb_terminal/account/txt/test_account_controller/test_print_help.txt
@@ -5,7 +5,7 @@ Cloud storage of routines:
     list               List available routines
     upload             Upload routine
     download           Download routine
-    delete             Delete routine
+    delete             Delete routine stored in cloud
 
 OpenBB Personal Access Token:
     generate           Generate a Personal Access Token
