+++ tests/openbb_terminal/session/test_local_model.py
@@ -15,9 +15,6 @@ from openbb_terminal.core.models.user_model import (
 )
 from openbb_terminal.core.session import local_model
 
-# IMPORTATION INTERNAL
-from openbb_terminal.core.session.current_user import get_current_user
-
 TEST_SESSION = {
     "access_token": "test_token",
     "token_type": "bearer",
@@ -156,158 +153,3 @@ def test_remove_cli_history_file_exception():
 
     os_mock.path.isfile.assert_called_with(local_model.HIST_FILE_PATH)
     os_mock.remove.assert_called_with(local_model.HIST_FILE_PATH)
-
-
-# Configs to apply
-CONFIGS = {
-    "features_settings": {
-        "USE_WATERMARK": "False",
-        "TIMEZONE": "Europe/London",
-        "PLOT_DPI": "95",
-        "PLOT_HEIGHT_PERCENTAGE": "50.5",
-        "USER_DATA_DIRECTORY": "some/path/to/user/data",
-    },
-    "features_keys": {
-        "API_KEY_ALPHAVANTAGE": "test_av",
-        "API_FRED_KEY": "test_fred",
-    },
-}
-
-
-@pytest.mark.parametrize(
-    "exists",
-    [
-        False,
-        True,
-    ],
-)
-def test_get_routine(mocker, exists: bool, test_user):
-    file_name = "test_routine.openbb"
-    uuid = "test_uuid"
-    routine = "do something"
-    current_user = get_current_user()
-
-    test_user.profile.uuid = uuid
-    mocker.patch(
-        target="openbb_terminal.core.session.local_model.get_current_user",
-        return_value=test_user,
-    )
-
-    exists_mock = mocker.patch(
-        "openbb_terminal.core.session.local_model.os.path.exists", return_value=exists
-    )
-    open_mock = mocker.patch(
-        "openbb_terminal.core.session.local_model.open",
-        mock_open(read_data=json.dumps(routine)),
-    )
-
-    assert local_model.get_routine(file_name=file_name) == json.dumps(routine)
-
-    exists_mock.assert_called_with(
-        current_user.preferences.USER_ROUTINES_DIRECTORY / uuid / file_name
-    )
-    if exists:
-        open_mock.assert_called_with(
-            current_user.preferences.USER_ROUTINES_DIRECTORY / uuid / file_name
-        )
-    else:
-        open_mock.assert_called_with(
-            current_user.preferences.USER_ROUTINES_DIRECTORY / file_name
-        )
-
-
-def test_get_routine_exception(mocker, test_user):
-    file_name = "test_routine.openbb"
-    uuid = "test_uuid"
-    current_user = get_current_user()
-
-    test_user.profile.uuid = uuid
-    mocker.patch(
-        target="openbb_terminal.core.session.local_model.get_current_user",
-        return_value=test_user,
-    )
-    exists_mock = mocker.patch(
-        "openbb_terminal.core.session.local_model.os.path.exists"
-    )
-    open_mock = mocker.patch(
-        "openbb_terminal.core.session.local_model.open",
-        side_effect=Exception("test exception"),
-    )
-
-    assert local_model.get_routine(file_name=file_name) is None
-
-    exists_mock.assert_called_with(
-        current_user.preferences.USER_ROUTINES_DIRECTORY / uuid / file_name
-    )
-    open_mock.assert_called_with(
-        current_user.preferences.USER_ROUTINES_DIRECTORY / uuid / file_name
-    )
-
-
-@pytest.mark.parametrize(
-    "exists",
-    [
-        False,
-        True,
-    ],
-)
-def test_save_routine(mocker, exists: bool, test_user):
-    file_name = "test_routine.openbb"
-    uuid = "test_uuid"
-    routine = "do something"
-    current_user = get_current_user()
-
-    test_user.profile.uuid = uuid
-    mocker.patch(
-        target="openbb_terminal.core.session.local_model.get_current_user",
-        return_value=test_user,
-    )
-
-    exists_mock = mocker.patch(
-        "openbb_terminal.core.session.local_model.os.path.exists", return_value=exists
-    )
-    makedirs_mock = mocker.patch("openbb_terminal.core.session.local_model.os.makedirs")
-    open_mock = mocker.patch(
-        "openbb_terminal.core.session.local_model.open",
-    )
-
-    result = local_model.save_routine(file_name=file_name, routine=routine)
-
-    if exists:
-        assert result == "File already exists"
-        makedirs_mock.assert_not_called()
-    else:
-        assert (
-            result
-            == current_user.preferences.USER_ROUTINES_DIRECTORY / uuid / file_name
-        )
-        makedirs_mock.assert_called_once()
-        open_mock.assert_called_with(
-            current_user.preferences.USER_ROUTINES_DIRECTORY / uuid / file_name, "w"
-        )
-
-    assert exists_mock.call_count == 2
-
-
-def test_save_routine_exception(mocker, test_user):
-    file_name = "test_routine.openbb"
-    uuid = "test_uuid"
-    routine = "do something"
-
-    test_user.profile.uuid = uuid
-    mocker.patch(
-        target="openbb_terminal.core.session.local_model.get_current_user",
-        return_value=test_user,
-    )
-
-    mocker.patch(
-        "openbb_terminal.core.session.local_model.os.path.exists", return_value=False
-    )
-    mocker.patch("openbb_terminal.core.session.local_model.os.makedirs")
-    mocker.patch(
-        "openbb_terminal.core.session.local_model.open",
-        side_effect=Exception("test exception"),
-    )
-
-    result = local_model.save_routine(file_name=file_name, routine=routine)
-    assert result is None
